renamed server configuration: move getServerConfig to middleware and update imports.
